+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 - **Package Handler** Fix: Minor label got mistaken as major in a specific case
 - **Web Scraper** Fix: Limit Playwright to only use English as locale when crawling websites
+- **Changelog** Fix: In cases of major releases the changelog writer was confused with where to write the Arch package changelog and where to write the origin package changelog
 - **Changelog** Fix: In some cases where the arch package tag and the origin package tag weren't the same it could have confused the changelog writer
 - **Package Handler** Fix: A return case where it did not check if there already was package changelog stored and always returned None